No font size adjustment for paragraphs inside list items
Markdown footnotes are rendered as a paragraph element inside a list
item element. The styling for this looked odd, because the font size
was being adjusted twice, once by style for the paragraph element and
once for the style of the list item element.

I'm not sure this is the best fix, but it'll do for now.