Remove filter property from frontmatter
I believe this was used by Webby to apply multiple template rendering
engines to a source file, e.g. ERB and then Textile. However, in
Middleman this is achieved using file extensions, e.g.
*.html.textile.erb. There weren't any ERB tags in the files in this
commit and they already had the *.textile file extension, so nothing was
lost in making this change and this was confirmed when I did `middleman
build` in that all the files were "identical".

Note that this also fixed a bunch of warnings which were being generated
at build time:

    WARN -- : You are setting a key that conflicts with a built-in method Middleman::Util::EnhancedHash#filter defined in Hash. This can cause unexpected behavior when accessing the key as a property. You can still access the key via the #[] method.