Use nix package manager for dev environment & CI build
* A gemset.nix was generated by running bundix against Gemfile.lock.

* The outline of shell.nix was generated by using --init option when
running bundix; however I also added 'ruby' to the buildInputs list to
ensure the right version of Ruby is used within the nix shell.

* I also added 'nodejs' to the buildInputs, because the execjs gem
requires it.

* I previously upgraded the app to the latest Ruby v2.6 patch version
and upgraded bundler to match the bundler installed in that version;
shell.nix does specify Ruby v2.6, but not the patch version or the
version of bundler.

* Use latest Ubuntu GitHub virtual environment instead of ruby:2.6.6
container image. This is partly because we now get Ruby from the
nix-shell and partly because we need some OS packages which are included
in the GitHub virtual environment for the install-nix-action step to
work.

* It might be worth using https://github.com/cachix/cachix-action to
speed up the build.