Use nix package manager for dev environment & CI build
* gemset.nix was generated by running bundix against Gemfile.lock.

* outline of shell.nix was generated by using --init option when running
bundix; however I also added 'ruby' to the buildInputs list to ensure
the right version of ruby is used within the nix shell.

* I haven't included node.js yet, because it doesn't actually seem to be
needed even though the coffee-script & uglifier gem supposedly need it.

* I've had to upgrade the app to the latest Ruby v2.6 patch version and
upgrade bundler to match the bundler installed in that version;
shell.nix does specify Ruby v2.6, but not the patch version or the
version of bundler.

* It might be worth using https://github.com/cachix/cachix-action to
speed up the build.