Move stylesheets to standard Middleman directory
I originally chose to use the css_dir config setting to use the `style`
directory [1] in order to make comparing the Webby version of the site
with the Middleman version while I was making the transition. This is no
longer necessary and so it seems better to use the default directory.

[1]: 9ee16d547dd8e65db1bdcb4246b12af16202a4c0