Simplify CSS for paragraphs inside list items
Peviously the "li > p" selector only caught cases where the paragraph
was an immediate descendant of the list item. However, I'm pretty sure
the intention was to apply this to any paragraph nested inside a list
item. This also allows me to remove an extra style which I had
previously had to add for the list of webmentions.